+++ app/src/ui/multi-commit-operation/base-multi-commit-operation.tsx
@@ -7,10 +7,10 @@ import { getResolvedFiles } from '../../lib/status'
 import { ConflictState, IMultiCommitOperationState } from '../../lib/app-state'
 import { Branch } from '../../models/branch'
 import { MultiCommitOperationStepKind } from '../../models/multi-commit-operation'
-import { ConflictsDialog } from './conflicts-dialog'
-import { ConfirmAbortDialog } from './confirm-abort-dialog'
-import { ProgressDialog } from './progress-dialog'
-import { WarnForcePushDialog } from './warn-force-push-dialog'
+import { ConflictsDialog } from './dialog/conflicts-dialog'
+import { ConfirmAbortDialog } from './dialog/confirm-abort-dialog'
+import { ProgressDialog } from './dialog/progress-dialog'
+import { WarnForcePushDialog } from './dialog/warn-force-push-dialog'
 import { PopupType } from '../../models/popup'
 
 export interface IMultiCommitOperationProps {
