+++ app/src/ui/multi-commit-operation/dialog/confirm-abort-dialog.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
-import { Dialog, DialogContent, DialogFooter } from '../dialog'
-import { OkCancelButtonGroup } from '../dialog/ok-cancel-button-group'
+import { Dialog, DialogContent, DialogFooter } from '../../dialog'
+import { OkCancelButtonGroup } from '../../dialog/ok-cancel-button-group'
 
 interface IConfirmAbortDialogProps {
   /**
