+++ app/src/ui/rebase/rebase-flow.tsx
@@ -16,7 +16,7 @@ import { RebaseProgressDialog } from './progress-dialog'
 import { ConfirmAbortDialog } from './confirm-abort-dialog'
 import { getResolvedFiles } from '../../lib/status'
 import { WarnForcePushDialog } from './warn-force-push-dialog'
-import { ConflictsDialog } from '../multi-commit-operation/conflicts-dialog'
+import { ConflictsDialog } from '../multi-commit-operation/dialog/conflicts-dialog'
 import { IMultiCommitOperationProgress } from '../../models/progress'
 import { RebaseChooseBranchDialog } from '../multi-commit-operation/choose-branch/rebase-choose-branch-dialog'
 import { MultiCommitOperationKind } from '../../models/multi-commit-operation'
